pyproject.toml
[project]
name = "automatizador-rutinas-gym"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.2.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
]
